add useContactForm hook
