+++ components/contact/ContactForm.tsx
@@ -1,82 +1,17 @@
 'use client';
-import { useState } from 'react';
+import useContactForm from '@/hooks/useContactForm';
 import ContactLinks from './ContactLinks';
 import clsx from 'clsx';
 
 const ContactForm = ({ dark = false }: { dark?: boolean }) => {
-   const initialData = {
-      name: '',
-      email: '',
-      message: ''
-   };
-   const [formData, setFormData] = useState({ ...initialData });
-   const [isSending, setIsSending] = useState(false);
-   const [isSent, setIsSent] = useState(false);
-   const [errorData, setErrorData] = useState<ErrorData | null>(null);
-
-   const handleChange = (
-      e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-   ) => {
-      const { name, value } = e.target;
-      setFormData((prevData: typeof initialData) => ({
-         ...prevData,
-         [name]: value
-      }));
-   };
-
-   const handleSending = () => {
-      setIsSending(true);
-   };
-
-   const handleSuccess = () => {
-      setIsSending(false);
-      setIsSent(true);
-      setFormData(initialData);
-   };
-
-   interface ErrorData {
-      message?: string;
-      [key: string]: unknown;
-   }
-
-   const handleError = (error: ErrorData) => {
-      setIsSending(false);
-      setIsSent(false);
-      setErrorData(error);
-   };
-
-   interface FetchErrorData extends ErrorData {
-      message: string;
-   }
-
-   const handleSubmit = async (
-      e: React.FormEvent<HTMLFormElement>
-   ): Promise<void> => {
-      e.preventDefault();
-      handleSending();
-      try {
-         const response: Response = await fetch('/api/send-email', {
-            method: 'POST',
-            headers: {
-               'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(formData)
-         });
-
-         if (response.ok) {
-            handleSuccess();
-         } else {
-            console.info('bb ~ response:', response);
-            const errorData: FetchErrorData = {
-               message: await response.text()
-            };
-            handleError(errorData);
-         }
-      } catch (error) {
-         console.error('bb ~ error:', error);
-         handleError(error as ErrorData);
-      }
-   };
+   const {
+      formData,
+      isSending,
+      isSent,
+      errorData,
+      handleChange,
+      handleSubmit
+   } = useContactForm();
 
    const SendingComponent = () => (
       <div className="mt-8 flex flex-col items-center gap-4">
@@ -91,11 +26,11 @@ const ContactForm = ({ dark = false }: { dark?: boolean }) => {
       </div>
    );
 
-   interface ErrorComponentProps {
-      error: ErrorData;
-   }
-
-   const ErrorComponent = ({ error }: ErrorComponentProps) => (
+   const ErrorComponent = ({
+      error
+   }: {
+      error: Record<string, unknown> | null;
+   }) => (
       <div className="mt-8 flex flex-col items-center gap-4">
          <p>Error: {JSON.stringify(error)}</p>
          <ContactLinks />
@@ -159,8 +94,8 @@ const ContactForm = ({ dark = false }: { dark?: boolean }) => {
                   type="submit"
                   id="submitBtn"
                   className={clsx(
-                     'hover-delay hover-brightness mt-4 w-full rounded-md  p-4 text-center font-roboto-sans text-sm tracking-wide  md:w-40',
-                     dark ? 'dark-text bg-bb-teal' : 'text-bb-teal bg-bb-gray'
+                     'hover-delay hover-brightness mt-4 w-full rounded-md p-4 text-center font-roboto-sans text-sm tracking-wide md:w-40',
+                     dark ? 'dark-text bg-bb-teal' : 'bg-bb-gray text-bb-teal'
                   )}>
                   Send
                </button>
